add documentation to types. Move all types to types
